This should fix Quest sharing, MAKE SURE SHARING IS ENABLED IN DB! i kept failing because of this lol

+++ src/server/game/Entities/Player/Player.cpp
@@ -18174,14 +18174,15 @@ void Player::SendQuestConfirmAccept(const Quest* quest, Player* pReceiver)
     }
 }
 
-void Player::SendPushToPartyResponse(Player* player, uint32 msg)
+void Player::SendPushToPartyResponse(Player* player, uint8 msg)
 {
     if (player)
     {
         WorldPacket data(MSG_QUEST_PUSH_RESULT, (8 + 1));
         data << uint64(player->GetGUID());
         data << uint8(msg);                                 // valid values: 0-8
-        GetSession()->SendPacket(&data);
+		GetSession()->SendPacket(&data);
+		
         sLog->outDebug(LOG_FILTER_NETWORKIO, "WORLD: Sent MSG_QUEST_PUSH_RESULT");
     }
 }
